requirements.txt
asgiref==3.7.2
Django==3.2.24
pytz==2024.1
sqlparse==0.4.4

djangorestframework-simplejwt==5.3

django-filter